+++ internal/kafka/command_topic.go
@@ -5,6 +5,7 @@ import (
 	"os"
 	"os/signal"
 	"runtime"
+	"strings"
 	"time"
 
 	"github.com/spf13/cobra"
@@ -14,6 +15,7 @@ import (
 	"github.com/confluentinc/cli/v3/pkg/ccloudv2"
 	pcmd "github.com/confluentinc/cli/v3/pkg/cmd"
 	"github.com/confluentinc/cli/v3/pkg/config"
+	dynamicconfig "github.com/confluentinc/cli/v3/pkg/dynamic-config"
 	"github.com/confluentinc/cli/v3/pkg/errors"
 	"github.com/confluentinc/cli/v3/pkg/log"
 	"github.com/confluentinc/cli/v3/pkg/output"
@@ -38,28 +40,29 @@ func newTopicCommand(cfg *config.Config, prerunner pcmd.PreRunner) *cobra.Comman
 
 	c := &command{clientID: cfg.Version.ClientID}
 
-	if cfg.IsCloudLogin() {
+	if cfg.CheckIsCloudLoginOrOnPremLogin() != nil {
+		c.AuthenticatedCLICommand = &pcmd.AuthenticatedCLICommand{CLICommand: pcmd.NewAnonymousCLICommand(cmd, prerunner)}
+	} else if cfg.IsCloudLogin() {
 		c.AuthenticatedCLICommand = pcmd.NewAuthenticatedCLICommand(cmd, prerunner)
 
-		cmd.AddCommand(c.newConsumeCommand())
 		cmd.AddCommand(c.newCreateCommand())
 		cmd.AddCommand(c.newDeleteCommand())
 		cmd.AddCommand(c.newDescribeCommand())
 		cmd.AddCommand(c.newListCommand())
-		cmd.AddCommand(c.newProduceCommand())
 		cmd.AddCommand(c.newUpdateCommand())
 	} else {
 		c.AuthenticatedCLICommand = pcmd.NewAuthenticatedWithMDSCLICommand(cmd, prerunner)
 		c.PersistentPreRunE = prerunner.InitializeOnPremKafkaRest(c.AuthenticatedCLICommand)
 
-		cmd.AddCommand(c.newConsumeCommandOnPrem())
 		cmd.AddCommand(c.newCreateCommandOnPrem())
 		cmd.AddCommand(c.newDeleteCommandOnPrem())
 		cmd.AddCommand(c.newDescribeCommandOnPrem())
 		cmd.AddCommand(c.newListCommandOnPrem())
-		cmd.AddCommand(c.newProduceCommandOnPrem())
 		cmd.AddCommand(c.newUpdateCommandOnPrem())
 	}
+
+	cmd.AddCommand(c.newConsumeCommand())
+	cmd.AddCommand(c.newProduceCommand())
 
 	return cmd
 }
@@ -139,6 +142,40 @@ func (c *command) provisioningClusterCheck(lkc string) error {
 	if cluster.Status.Phase == ccloudv2.StatusProvisioning {
 		return fmt.Errorf(errors.KafkaRestProvisioningErrorMsg, lkc)
 	}
+	return nil
+}
+
+func (c *command) prepareAnonymousContext(cmd *cobra.Command) error {
+	kafkaBootstrap, err := cmd.Flags().GetString("kafka-bootstrap")
+	if err != nil {
+		return err
+	}
+
+	platform := &config.Platform{
+		Server: kafkaBootstrap,
+		Name:   strings.TrimPrefix(kafkaBootstrap, "https://"),
+	}
+
+	kafkaClusterCfg := &config.KafkaClusterConfig{
+		ID:        "anonymous-id",
+		Name:      "anonymous-cluster",
+		Bootstrap: kafkaBootstrap,
+		APIKeys:   map[string]*config.APIKeyPair{},
+	}
+	kafkaClusters := map[string]*config.KafkaClusterConfig{kafkaClusterCfg.ID: kafkaClusterCfg}
+
+	ctx := &config.Context{Platform: platform}
+	kafkaClusterContext := &config.KafkaClusterContext{
+		EnvContext:          false,
+		ActiveKafkaCluster:  kafkaClusterCfg.ID,
+		KafkaClusterConfigs: kafkaClusters,
+		Context:             ctx,
+	}
+	ctx.KafkaClusterContext = kafkaClusterContext
+
+	dynamicContext := &dynamicconfig.DynamicContext{Context: ctx}
+	c.Context = dynamicContext
+
 	return nil
 }
