+++ internal/kafka/command_topic_consume.go
@@ -28,7 +28,6 @@ func (c *command) newConsumeCommand() *cobra.Command {
 		Args:              cobra.ExactArgs(1),
 		ValidArgsFunction: pcmd.NewValidArgsFunction(c.validArgs),
 		RunE:              c.consume,
-		Annotations:       map[string]string{pcmd.RunRequirement: pcmd.RequireCloudLogin},
 		Example: examples.BuildExampleString(
 			examples.Example{
 				Text: `Consume items from topic "my-topic" and press "Ctrl-C" to exit.`,
@@ -50,8 +49,11 @@ func (c *command) newConsumeCommand() *cobra.Command {
 	cmd.Flags().Bool("timestamp", false, "Print message timestamp in milliseconds.")
 	cmd.Flags().StringSlice("config", nil, `A comma-separated list of configuration overrides ("key=value") for the consumer client.`)
 	pcmd.AddConsumerConfigFileFlag(cmd)
-	cmd.Flags().String("schema-registry-context", "", "The Schema Registry context under which to look up schema ID.")
 	cmd.Flags().String("schema-registry-endpoint", "", "Endpoint for Schema Registry cluster.")
+
+	// cloud-only flags
+	cmd.Flags().String("kafka-bootstrap", "", "Bootstrap URL for Kafka cluster.")
+	cmd.Flags().String("schema-registry-context", "", "The Schema Registry context under which to look up schema ID.")
 	cmd.Flags().String("schema-registry-api-key", "", "Schema registry API key.")
 	cmd.Flags().String("schema-registry-api-secret", "", "Schema registry API secret.")
 	pcmd.AddApiKeyFlag(cmd, c.AuthenticatedCLICommand)
@@ -59,6 +61,11 @@ func (c *command) newConsumeCommand() *cobra.Command {
 	pcmd.AddClusterFlag(cmd, c.AuthenticatedCLICommand)
 	pcmd.AddContextFlag(cmd, c.CLICommand)
 	pcmd.AddEnvironmentFlag(cmd, c.AuthenticatedCLICommand)
+
+	// on-prem only flags
+	cmd.Flags().AddFlagSet(pcmd.OnPremAuthenticationSet())
+	pcmd.AddProtocolFlag(cmd)
+	pcmd.AddMechanismFlag(cmd, c.AuthenticatedCLICommand)
 
 	cobra.CheckErr(cmd.MarkFlagFilename("config-file", "avsc", "json"))
 
@@ -69,6 +76,21 @@ func (c *command) newConsumeCommand() *cobra.Command {
 }
 
 func (c *command) consume(cmd *cobra.Command, args []string) error {
+	if c.Context == nil || c.Context.State == nil {
+		cobra.CheckErr(cmd.MarkFlagRequired("kafka-bootstrap"))
+		if err := c.prepareAnonymousContext(cmd); err != nil {
+			return err
+		}
+		return c.consumeCloud(cmd, args)
+	} else if c.Context.Config.IsCloudLogin() {
+		return c.consumeCloud(cmd, args)
+	} else {
+		cobra.CheckErr(cmd.MarkFlagRequired("bootstrap"))
+		return c.consumeOnPrem(cmd, args)
+	}
+}
+
+func (c *command) consumeCloud(cmd *cobra.Command, args []string) error {
 	topic := args[0]
 
 	cluster, err := c.Context.GetKafkaClusterForCommand(c.V2Client)
@@ -218,6 +240,124 @@ func (c *command) consume(cmd *cobra.Command, args []string) error {
 			Timestamp:   timestamp,
 			Delimiter:   delimiter,
 			SchemaPath:  schemaPath,
+		},
+	}
+	return RunConsumer(consumer, groupHandler)
+}
+
+func (c *command) consumeOnPrem(cmd *cobra.Command, args []string) error {
+	// add print-offset
+	printKey, err := cmd.Flags().GetBool("print-key")
+	if err != nil {
+		return err
+	}
+
+	fullHeader, err := cmd.Flags().GetBool("full-header")
+	if err != nil {
+		return err
+	}
+
+	timestamp, err := cmd.Flags().GetBool("timestamp")
+	if err != nil {
+		return err
+	}
+
+	delimiter, err := cmd.Flags().GetString("delimiter")
+	if err != nil {
+		return err
+	}
+
+	keyFormat, err := cmd.Flags().GetString("key-format")
+	if err != nil {
+		return err
+	}
+
+	valueFormat, err := cmd.Flags().GetString("value-format")
+	if err != nil {
+		return err
+	}
+
+	configFile, err := cmd.Flags().GetString("config-file")
+	if err != nil {
+		return err
+	}
+	config, err := cmd.Flags().GetStringSlice("config")
+	if err != nil {
+		return err
+	}
+
+	consumer, err := newOnPremConsumer(cmd, c.clientID, configFile, config)
+	if err != nil {
+		return errors.NewErrorWithSuggestions(
+			fmt.Sprintf(errors.FailedToCreateConsumerErrorMsg, err),
+			errors.OnPremConfigGuideSuggestions,
+		)
+	}
+	log.CliLogger.Tracef("Create consumer succeeded")
+
+	if err := c.refreshOAuthBearerToken(cmd, consumer); err != nil {
+		return err
+	}
+
+	adminClient, err := ckafka.NewAdminClientFromConsumer(consumer)
+	if err != nil {
+		return fmt.Errorf(errors.FailedToCreateAdminClientErrorMsg, err)
+	}
+	defer adminClient.Close()
+
+	topicName := args[0]
+	if err := ValidateTopic(adminClient, topicName); err != nil {
+		return err
+	}
+
+	offset, err := GetOffsetWithFallback(cmd)
+	if err != nil {
+		return err
+	}
+
+	partition, err := cmd.Flags().GetInt32("partition")
+	if err != nil {
+		return err
+	}
+	partitionFilter := PartitionFilter{
+		Changed: cmd.Flags().Changed("partition"),
+		Index:   partition,
+	}
+
+	rebalanceCallback := GetRebalanceCallback(offset, partitionFilter)
+	if err := consumer.Subscribe(topicName, rebalanceCallback); err != nil {
+		return err
+	}
+
+	output.ErrPrintln(c.Config.EnableColor, errors.StartingConsumerMsg)
+
+	var srClient *schemaregistry.Client
+	if slices.Contains(serdes.SchemaBasedFormats, valueFormat) {
+		srClient, err = c.GetSchemaRegistryClient(cmd)
+		if err != nil {
+			return err
+		}
+	}
+
+	dir, err := sr.CreateTempDir()
+	if err != nil {
+		return err
+	}
+	defer func() {
+		_ = os.RemoveAll(dir)
+	}()
+
+	groupHandler := &GroupHandler{
+		SrClient:    srClient,
+		KeyFormat:   keyFormat,
+		ValueFormat: valueFormat,
+		Out:         cmd.OutOrStdout(),
+		Properties: ConsumerProperties{
+			PrintKey:   printKey,
+			FullHeader: fullHeader,
+			Timestamp:  timestamp,
+			Delimiter:  delimiter,
+			SchemaPath: dir,
 		},
 	}
 	return RunConsumer(consumer, groupHandler)
